+++ module-common/src/test/java/kichenpos/common/ApplicationTest.java
@@ -1,0 +1,11 @@
+package kichenpos.common;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTest {
+    @Test
+    void contextLoads() {
+    }
+}
