+++ module-menu/src/main/java/kichenpos/menu/Application.java
@@ -1,0 +1,11 @@
+package kichenpos.menu;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
