+++ module-menu/src/test/java/kichenpos/menu/domain/MenuProductTest.java
@@ -1,0 +1,23 @@
+package kichenpos.menu.domain;
+
+import kichenpos.common.domain.Price;
+import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class MenuProductTest {
+
+    @Test
+    void 메뉴_상품의_가격을_계산한다() {
+        // given
+        MenuProduct menuProduct = new MenuProduct(new Product("치킨", new Price(BigDecimal.valueOf(5000))), 5);
+
+        // when
+        Price result = menuProduct.price();
+
+        // then
+        assertThat(result).isEqualTo(new Price(BigDecimal.valueOf(25000)));
+    }
+}
