+++ module-order/src/main/java/kichenpos/order/Application.java
@@ -1,0 +1,11 @@
+package kichenpos.order;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
