+++ module-table/src/main/java/kichenpos/table/Application.java
@@ -1,0 +1,11 @@
+package kichenpos.table;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
